Adding Vuex as Global State mangament
Even if this is a small project, having a statement in the project will help and show I know the stuff needed to work as a team. I don't want to deal with annoying data buses as well as having to send data down from child to parent or parent to child unless it is props down to read-only components.